♻️ NavLinksコンポーネントを作成
