+++ src/pages-layout/ContentLayout/ContentLayout.tsx
@@ -1,16 +1,6 @@
 import { Burger, CloseButton, Drawer } from '@mantine/core'
 import { useDisclosure } from '@mantine/hooks'
-import {
-  IconArrowBackUp,
-  IconChartLine,
-  IconLock,
-  IconNotes,
-  IconShare,
-} from '@tabler/icons'
-import { useRouter } from 'next/router'
-import { ReactElement, useEffect } from 'react'
-import { getPath } from 'src/lib/const'
-import { ActiveLink } from 'src/lib/next'
+import { NavLinks } from './NavLinks'
 import { ErrorBoundary } from 'src/pages-layout/ErrorBoundary'
 import { HeaderWrapper } from 'src/pages-layout/HeaderWrapper'
 
@@ -29,31 +19,8 @@ export const ContentLayout = (page: ReactElement) => {
 
 const Header = () => {
   const router = useRouter()
-  const planId = router.query.planId
-  const [opened, handlers] = useDisclosure(false)
-
-  const LINKS = [
-    {
-      href: getPath('PLAN', String(planId)),
-      label: '計画表',
-      icon: <IconNotes size={28} stroke={1.7} />,
-    },
-    {
-      href: getPath('PREF_NEWS', String(planId)),
-      label: '旅先の情報',
-      icon: <IconChartLine size={28} stroke={1.7} />,
-    },
-    {
-      href: getPath('FORGOT_PASSWORD', String(planId)),
-      label: 'パスワード再設定',
-      icon: <IconLock size={28} stroke={1.7} />,
-    },
-    {
-      href: getPath('INDEX'),
-      label: 'トップへ戻る',
-      icon: <IconArrowBackUp size={28} stroke={1.7} />,
-    },
-  ] as const
+  const [drawerOpened, drawerHandlers] = useDisclosure(false)
+  const [modalOpened, modalHandlers] = useDisclosure(true)
 
   // パスの変更を検知して Drawer を閉じる処理
   useEffect(() => {
@@ -82,31 +49,7 @@ const Header = () => {
           Freely
         </div>
         <hr className='mx-2 h-[1px] border-0 bg-dark-100' />
-        <div className='mt-10 ml-4 space-y-4'>
-          {LINKS.map(({ href, label, icon }) => {
-            return (
-              <ActiveLink href={href} passHref key={href}>
-                {(isActive) => {
-                  return (
-                    <a
-                      className={`flex items-center no-underline ${
-                        isActive ? 'text-main-500' : 'text-dark-500'
-                      }`}
-                    >
-                      <div
-                        className={`mr-4 h-10 w-1 rounded-sm ${
-                          isActive ? 'bg-main-500' : 'bg-white'
-                        }`}
-                      ></div>
-                      {icon}
-                      <div className='ml-2'>{label}</div>
-                    </a>
-                  )
-                }}
-              </ActiveLink>
-            )
-          })}
-        </div>
+        <NavLinks />
       </Drawer>
       <div className='flex items-center gap-8'>
         <Burger
